feat: 앱라우터 error페이지용 에러 디스플레이 컴포넌트 추가
