+++ src/app/post/[slug]/error.tsx
@@ -1,48 +1,15 @@
 'use client';
 
-import ErrorPageContainer from '@/ui/error/ErrorPageContainer';
-import ErrorMessage from '@/ui/error/ErrorMessage';
-import Button from '@/shared/ui/Button';
-import { useRouter } from 'next/navigation';
-import { useResetRecoilState } from 'recoil';
-import { activeMainBoardTabStore } from '@/store/ActiveMainBoardTabStateStore';
-import {
-  selectedPositionState,
-  selectedTechStackState,
-} from '@/features/post/public/store/PostSearchStateStore';
-import { useEffect } from 'react';
+import ErrorPageDisplay from '@/components/error/ErrorPageDisplay';
 
-export default function Error({
+const PostDetailPageError = ({
   error,
   reset,
 }: {
   error: Error & { digest?: string };
   reset: () => void;
-}) {
-  const router = useRouter();
-  const resetActiveBoardTab = useResetRecoilState(activeMainBoardTabStore);
-  const resetSelectedTechStacks = useResetRecoilState(selectedTechStackState);
-  const resetSelectedPosition = useResetRecoilState(selectedPositionState);
+}) => {
+  return <ErrorPageDisplay error={error} reset={reset} />;
+};
 
-  const goHome = () => {
-    resetActiveBoardTab();
-    resetSelectedTechStacks();
-    resetSelectedPosition();
-
-    router.push('/');
-  };
-
-  useEffect(() => {
-    console.error(error);
-  }, [error]);
-
-  return (
-    <ErrorPageContainer>
-      <ErrorMessage>에러가 발생했습니다.</ErrorMessage>
-      <div className='min-h-[80px] flex items-center space-x-2'>
-        <Button onClick={() => reset()}>재시도</Button>
-        <Button onClick={goHome}>홈으로</Button>
-      </div>
-    </ErrorPageContainer>
-  );
-}
+export default PostDetailPageError;
